+++ api_v1/attraction.py
@@ -2,8 +2,15 @@ from flask import make_response, request, redirect
 import json
 
 from model.db import Mydb
-from model.data_formatter import attractionsFormatter
 from . import api
+
+# 將由資料庫取得的景點(attractions)，整理成dict格式
+def attractionsFormatter(result):
+    cols = ["id", "name", "category", "description", "address", "transport", "mrt", "latitude", "longitude", "images"]
+    data_dict = dict(zip(cols, result))
+    images = result[9].split(" ")
+    data_dict["images"] = images[:-1]
+    return data_dict
 
 # 取得景點資料列表api
 @api.route("/attractions")
